Splitting the declaration and implementation of methods into separate files
This allows you to re-include the library in different executables
